Time: 3 ms (96.7%), Space: 115.9 MB (14.42%) - LeetHub

+++ 2348-number-of-zero-filled-subarrays/2348-number-of-zero-filled-subarrays.java
@@ -1,7 +1,7 @@
 class Solution {
     public long zeroFilledSubarray(int[] nums) {
         long ans = 0;
-        int k = 0;
+        long k = 0;
         for(int i=0;i<nums.length;i++){
             int num = nums[i];
             if(num == 0) {
